requirements.txt
numpy
scipy
tqdm
lpips
mmcv
imageio
imageio-ffmpeg
opencv-python
ipdb
torch_scatter
einops
matplotlib
pytorch_msssim
bitstring
bitarray
bidict
ninja
pyqt5
